in1k/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.15.0=pypi_0
aiohttp=3.7.4.post0=pypi_0
async-timeout=3.0.1=pypi_0
attrs=21.2.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.9.30=h06a4308_1
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py39h06a4308_0
chardet=4.0.0=pypi_0
charset-normalizer=2.0.7=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
ffmpeg=4.3=hf484d3e_0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.10.1=pypi_0
future=0.18.2=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.3.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.41.0=pypi_0
idna=3.3=pypi_0
imageio=2.10.2=pypi_0
imageio-ffmpeg=0.4.5=pypi_0
intel-openmp=2021.3.0=h06a4308_3350
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
kiwisolver=1.3.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
lightning-bolts=0.4.0=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=pypi_0
matplotlib=3.4.3=pypi_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
moviepy=1.0.3=pypi_0
multidict=5.2.0=pypi_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
numpy=1.21.2=py39h20f2e39_0
numpy-base=1.21.2=py39h79a1101_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.0=hd408876_0
openssl=1.1.1l=h7f8727e_0
pandas=1.3.4=pypi_0
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
proglog=0.1.9=pypi_0
protobuf=3.19.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pydeprecate=0.3.1=pypi_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pypi_0
pytorch=1.10.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.4.9=pypi_0
pytorch-lightning-bolts=0.3.2=pypi_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pypi_0
pyyaml=6.0=pypi_0
readline=8.1=h27cfd23_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
setuptools=58.0.4=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
timm=0.4.12=pypi_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.10.0=py39_cu113
torchmetrics=0.5.1=pypi_0
torchvision=0.11.1=py39_cu113
tqdm=4.62.3=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021a=h5d7bf9c_0
urllib3=1.26.7=pypi_0
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h7b6447c_0
yarl=1.7.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
